begin migrating the site
So far I have migrated about 60% of all the pages and came up with basic
designs for about 40% of them. However many of the more important pages
either do not have a design or have a temporarily a crappy one.

Anyways, if you want to use media queries, use the @media mixin provided
by Neat to create it. This is important, as we need separatable media
queries organization that get included together at the end
(performance).