begin a rather crappy implementation of a gallery
So right now the gallery is rather crappy. What it does is gets some
Picasa JSON data, does stuff with it, and spits out a crappy gallery. It
basically works, but it is totally unoptimized. Not to mention the code
is still messy...

Maybe someone could write a jekyll-picasa plugin...